+++ pyproject.toml
@@ -28,12 +28,10 @@ dependencies = [
     "zarr",
     "numcodecs",
     "dask",
-    "matplotlib",
     "pydantic>=2.11.8",
     "pyproj",
     "pyaml_env",
     "pyresample",
-    "h5netcdf",
     "xarray-tensorstore==0.1.5",
     "zarr>=3",
 ]
@@ -56,8 +54,7 @@ repository = "https://github.com/openclimatefix/ocf-data-sampler"
 license-files = []
 
 [tool.setuptools.packages.find]
-where = ["."]
-exclude = ["test*"]
+where = ["src"]
 
 [tool.setuptools.package-data]
 "*" = ["*.csv"]
@@ -95,7 +92,8 @@ ignore_missing_imports = true
 [tool.ruff]
 line-length = 100
 indent-width = 4
-exclude = ["__init__.py"]
+force-exclude = true
+extend-exclude = ["src/ocf_data_sampler/features/solar.py"]
 
 [tool.ruff.lint]
 select = [
@@ -125,7 +123,8 @@ ignore = [
 ]
 
 [tool.ruff.lint.per-file-ignores]
-"test*" = ["D", "ANN", "S101"]
+"tests/**" = ["D", "ANN", "S101"]
+"__init__.py" = ["F401", "D104"]
 
 [tool.ruff.lint.pydocstyle]
 convention = "google"
