+++ src/ocf_data_sampler/config/__init__.py
@@ -1,5 +1,5 @@
-"""Configuration model"""
+"""Configuration model."""
 
+from ocf_data_sampler.config.load import load_yaml_configuration
 from ocf_data_sampler.config.model import Configuration, InputData
 from ocf_data_sampler.config.save import save_yaml_configuration
-from ocf_data_sampler.config.load import load_yaml_configuration
