+++ src/ocf_data_sampler/features/solar.py
@@ -1,4 +1,52 @@
-"""Module for calculating solar position."""
+"""Solar azimuth and elevation via an optimised port of pvlib's ephemeris algorithm.
+
+This module contains a copy of the `ephemeris()` function from the pvlib-python
+project [1], adapted for speed.
+
+Modifications copyright 2026 Open Climate Fix. Licensed under the MIT License.
+
+Modifications from the original include:
+- Accepts numpy datetime64 arrays directly (no pandas DatetimeIndex)
+- No atmospheric refraction correction (returns true elevation)
+- Time components computed via ocf_data_sampler.common.time_utils
+
+References:
+    [1] https://github.com/pvlib/pvlib-python/blob/main/pvlib/solarposition.py
+"""
+# The following notice is retained from pvlib-python as required by its license:
+#
+# BSD 3-Clause License
+#
+# Copyright (c) 2023 pvlib python Contributors
+# Copyright (c) 2014 PVLIB python Development Team
+# Copyright (c) 2013 Sandia National Laboratories
+#
+# All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without modification,
+# are permitted provided that the following conditions are met:
+#
+#   Redistributions of source code must retain the above copyright notice, this
+#   list of conditions and the following disclaimer.
+#
+#   Redistributions in binary form must reproduce the above copyright notice, this
+#   list of conditions and the following disclaimer in the documentation and/or
+#   other materials provided with the distribution.
+#
+#   Neither the name of the copyright holder nor the names of its
+#   contributors may be used to endorse or promote products derived from
+#   this software without specific prior written permission.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+# ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+# DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR
+# ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
+# ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 import numpy as np
 from numpy.typing import NDArray
