+++ src/ocf_data_sampler/load/nwp.py
@@ -1,20 +1,97 @@
 """Module for opening NWP data."""
 
 
-from typing import TYPE_CHECKING
+from glob import glob
+from typing import TYPE_CHECKING, Literal
 
 import numpy as np
 import xarray as xr
 
 from ocf_data_sampler.common.indexing import assert_values_unique_increasing
-from ocf_data_sampler.load.utils import (
+from ocf_data_sampler.common.xr_tensorstore import open_zarr, open_zarrs
+from ocf_data_sampler.load.conventions import (
     get_xr_data_array_from_xr_dataset,
     make_spatial_coords_increasing,
-    open_zarr_paths,
 )
 
 if TYPE_CHECKING:
     from collections.abc import Callable
+
+
+def open_zarr_paths(
+    zarr_path: str | list[str],
+    time_dim: str,
+    backend: Literal["dask", "tensorstore"],
+    public: bool = False,
+) -> xr.Dataset:
+    """Opens the NWP data.
+
+    Args:
+        zarr_path: Path to the zarr(s) to open
+        time_dim: Name of the time dimension
+        public: Whether the data is public or private. Only available for the dask backend.
+        backend: The xarray backend to use.
+
+    Returns:
+        The opened Xarray Dataset
+    """
+    if backend not in ["dask", "tensorstore"]:
+        raise ValueError(
+            f"Unsupported backend: {backend}. Supported backends are 'dask' and 'tensorstore'.",
+        )
+
+    if public and backend == "tensorstore":
+        raise ValueError("Public data is only supported with the 'dask' backend.")
+
+    if backend == "tensorstore":
+        ds = _tensorstore_open_zarr_paths(zarr_path, time_dim)
+
+    elif backend == "dask":
+        ds = _dask_open_zarr_paths(zarr_path, time_dim, public)
+
+    return ds
+
+
+def _dask_open_zarr_paths(zarr_path: str | list[str], time_dim: str, public: bool) -> xr.Dataset:
+    general_kwargs = {
+        "engine": "zarr",
+        "chunks": "auto",
+        "decode_timedelta": True,
+    }
+
+    if public:
+        # note this only works for s3 zarr paths at the moment
+        general_kwargs["storage_options"] = {"anon": True}
+
+    if isinstance(zarr_path, list | tuple) or "*" in str(zarr_path):  # Multi-file dataset
+        ds = xr.open_mfdataset(
+            zarr_path,
+            concat_dim=time_dim,
+            combine="nested",
+            coords="different",
+            compat="no_conflicts",
+            **general_kwargs,
+        ).sortby(time_dim)
+    else:
+        ds = xr.open_dataset(
+            zarr_path,
+            consolidated=True,
+            mode="r",
+            **general_kwargs,
+        )
+    return ds
+
+
+def _tensorstore_open_zarr_paths(zarr_path: str | list[str], time_dim: str) -> xr.Dataset:
+
+    if "*" in str(zarr_path):
+        zarr_path = sorted(glob(zarr_path))
+
+    if isinstance(zarr_path, list | tuple):
+        ds = open_zarrs(zarr_path, concat_dim=time_dim, data_source="nwp").sortby(time_dim)
+    else:
+        ds = open_zarr(zarr_path)
+    return ds
 
 
 def _validate_nwp_data(data_array: xr.DataArray, provider: str) -> None:
