+++ tests/common/test_tensorstore.py
@@ -1,7 +1,6 @@
 import pandas as pd
 import pytest
 import xarray as xr
-
 from ocf_data_sampler.common.tensorstore import open_zarr, open_zarrs
