+++ tests/datasets/pvnet/test_materialise.py
@@ -1,8 +1,8 @@
 import dask.array
 import numpy as np
 import xarray as xr
-
 from ocf_data_sampler.common.tensorstore import open_zarr
+
 from ocf_data_sampler.datasets.pvnet.materialise import load, load_data_dict
